requirements.txt
pdfminer.six
kokoro
soundfile
pydub